requirements.txt
Django>=4.2,<5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
drf-spectacular==0.27.0
drf-yasg==1.21.7
Pillow>=10.0.0
python-decouple==3.8
geopy==2.4.1
django-filter==23.5
python-dateutil==2.8.2
pytz==2024.1
celery==5.3.6
redis==5.0.1
requests==2.31.0
pydantic-core>=2.20

# GeoDjango dependencies (installer selon votre OS)
# Pour Ubuntu/Debian: apt-get install gdal-bin libgdal-dev libgeos-dev libproj-dev
# Pour macOS: brew install gdal geos proj

# Base de données spatiale
# spatialite sera installé automatiquement avec Django

# Utilitaires
faker==22.0.0
python-magic==0.4.27

# Tests
pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0

# Production (optionnel)
gunicorn==21.2.0
whitenoise==6.6.0
setuptools==68.0.0
